Hopper compiler fixes
Compiler:
- tricky compiler bug (incorrect cast of 'middle' argument)
- additional obvious automatic upcasts (byte -> +int, int, +int -> uint, -int to int)
Windows Runtime:
- named pipe work for 'COM0' tweaked more

+++ Source/Compiler/DASM.hs
@@ -133,7 +133,7 @@ program DASM
             <uint> jumpTargets;
             <uint> jixLabels;
             
-            content = Instructions.Disassemble(code, ref address, startAddress, ref jumpTargets, ref jixLabels, false);
+            content = Instructions.Disassemble(code, ref address, long(startAddress), ref jumpTargets, ref jixLabels, false);
             instructionCount++;
             String.Build(ref content, char(0x0A));
             hasmFile.Append(content);
